split up the source and target layer names
The source gives a full docker path, but for oci source we only use the last
part of the image name.  I.e. for docker://zothub.local/puzzleos/hostfs:1.0,
we only use 'host'.  The rest is to give an indication of where
the layer came from. (For remote zot, we will append 'puzzleos/hostfs:1.0'
to the provided baseUrl)

The Name is the target name, i.e. 'hostfs'.

The Fullname is the zot path - e.g. puzzleos/hostfs.

The Version is the version we'll want to call it on the installed system.
While hostfs may be at version 1.0, we may want to tie the installed
image version to something else on the system and call it 1.0.1.

Signed-off-by: Serge Hallyn <serge@hallyn.com>