Separate system manifest from install manifests
When we first install a system from an install.yaml, the system
manifest simply contains all the targets in the install.yaml.
However, if we do an update (not yet implemented) from a partial
manifest, then the resulting system manifest may contain targets
from multiple install manifests.

The install manifests are signed by the distributor, so we must
keep those and their signatures.

Therefore, upon install (and update), we build a 'system manifest',
which simply lists the names of the targets to run, and pointers
to the install.yamls from which they came.  Each install.yaml is
renamed to (its sha256sum).yaml, as are its .pem and .yaml.signed.

Signed-off-by: Serge Hallyn <serge@hallyn.com>