Test and fix re-activate for fs-only
When doing mos activate on an fs-only layer that's already
running, we were failing because the 'stop' action was not
being done.  In order to unmount the $chroot/mnt/atom/layername,
we have to pass the RootDir from mos to the atomfsstorage,
so atomfsstorage can know to umount just based on the layername.

Signed-off-by: Serge Hallyn <serge@hallyn.com>