implement mosb iso build
It takes a bare version of a install.yaml, consisting only of
targets, something like:

targets:
  - service_name: hostfs
    imagepath: oci:zothub:busybox-squashfs
    version: 1.0.0
    service_type: hostfs
    nsgroup: ""
    network:
      type: host
  - service_name: hostfstarget
    imagepath: docker://zothub.io/tools/busybox:stable
    version: 1.0.0
    service_type: fs-only
    nsgroup: ""
    network:
      type: none

It will rewrite the imagepath to reflect what to expect the
installed image to look like on the host.  It will create an
ISO containing a full install.yaml, the specified manifest
signing certificate and signed manifest, and an oci layout
with each specified target installed referenced by the
target.service_name as the image name.

Eventually we'll want this to also create an EFI partition
with the appropriate signed UKI and shim, to support simple
installation by ISO boot.

Signed-off-by: Serge Hallyn <serge@hallyn.com>