Add build tag containers_image_openpgp to drop libgpgme deps.
The net change of this commit is to drop the following C-go
dependencies from the 'mosctl' binary:

  * libassuan.so.0
  * libgpg-error.so.0
  * libgpgme.so.11

The net change on size of the binary is negligible, it is
reduced by ~200kb of a 28mb binary.

Per https://github.com/containers/image/blob/main/README.md

> containers_image_openpgp: Use a Golang-only OpenPGP
> implementation for signature verification instead of
> the default cgo/gpgme-based implementation; the primary
> downside is that creating new signatures with the
> Golang-only implementation is not supported.

The build still requires installation of libgpgme-dev,
but the built binary does not link to it.
Without it, build fails with:

  Error: ../../../go/pkg/mod/github.com/
      mtrmac/gpgme@v0.1.2/data.go:4:11:
     fatal error: gpgme.h: No such file or directory

Signed-off-by: Scott Moser <smoser@brickies.net>