Set a 20 minute timeout on 'make test'
Using 'timeout' was preferable to me compared to
jobs.<job_id>.timeout-minutes.  Testing showed that putting a
'timeout-minutes' on a jobs.<job_id>.steps.<step> did work, but
I liked the control of 'timeout'.  I think the github action
was killed with SIGKILL as no additional output of bats was shown.