cleanup towards 'make test' working. (#30)
* Fix Makefile so mosb and mosctl get built.

The default target had become 'zot'. That meant mosb
or mosctl were not built. Reorder targets and add zot and oras
to the 'all' target.

* Cleanup / Workflow changes towards a working 'make test'.

There are a lot of changes to the workflow file here that
ultimately result in being closer to getting 'make test'
to pass on github workflow.

Fixes:
 * Drop the 'ref: main' from checkout@v3 - this was causing us
   to actually build main even in someone's pull request.
 * add ~/bin to PATH via GITHUB_ENV - doing it once means it
   is there for all sections.
 * install packages for 'make test' to succeed:
    * libarchive-tools for bsdtar
    * squashfuse
    * umoci
 * install skopeo from project-machine/tools release,
   I think the snap was causing the newuidmap/newgidmap issues due
   to apparmor and the setuid on those files.
 * provide git with author and committer via environment vars.
   GIT_COMMITTER_AUTHOR / GIT_COMMITTER_EMAIL and
   GIT_AUTHOR_NAME / GIT_AUTHOR_EMAIL
	export GIT_AUTHOR_NAME="$me" GIT_COMMITTER_NAME="$me"

Cleanup:
 * do not install golang via apt, as we're using it via setup-go@v3.
 * drop some debug-like statements.

* Set a 20 minute timeout on 'make test'

Using 'timeout' was preferable to me compared to
jobs.<job_id>.timeout-minutes.  Testing showed that putting a
'timeout-minutes' on a jobs.<job_id>.steps.<step> did work, but
I liked the control of 'timeout'.  I think the github action
was killed with SIGKILL as no additional output of bats was shown.

* Set up lxc.conf, lxc test changes

Basically:
 * Write a lxc.conf for the running user so they can execute
   unpriv lxc.
 * Put a timeout on 'lxc-wait' calls
 * Remove one use of expect, replacing it with '-y' to
   add-apt-repository.
 * make tests/create-test-container.bash run with '-e' so
   it fails fast.