simplify verified remote layer mount
We used to use "soci", a specialized OCI layer containing
an install manifest, its signature, and the cert needed
to verify the signature.  Since we now have an OCI artifact
based way to ship that, drop the soci generation and parsing
code.

Add a new 'mosctl mount'.  If no optional argument is specified,
then it will mount the specified target from the current manifest.
If an optional argument is specified, then it is expected to be
a url to a remote manifest, like 10.0.2.2:puzzleos/livecd:1.0 .
That manifest must have two referring artifacts, one vnd.machine.signature
and one vnd.machine.pubkeycrt.  The cert must be signed by our
manifest ca, and the signature must be verified with that cert.

This can be used to mount a livecd rootfs image from a remote zot
server.

Changes:

Use an oci artifact for install manifest, plus two artifacts which refer
to the install manifest artifact, one carrying the signature and the
other carrying the certificate for verifying the signature.

Change the 'install.yaml' to a 'install.json'.  Install.json is
used by machines to install.

Introduce an imports.yaml.  This is used by admins build install.json.
See formats.md for the difference.  In particular, the remote
distribution URL for a layer is specified in imports.yaml, but not
in install.json.

Drop Targets.Mounts.  We will want to provide that functionality,
but more finely suited for (a) providing (id-mapped) cross-container
mounts, and (b) providing persistent storage on a host for use by
containers.

Drop the 'mosb iso' and 'mosb oci' commands.

Drop 'lxc' tests.  We'll hook up with mcli to do VM tests.

Drop the target.ImagePath.  We name all target layers in the local
repository as "docker://localhost/mos:digest".  The digest is specified
in the Target struct.

go.mod: switch to newer containers/image, because the fix for copying
compressed squashfs images is STILL not in a release!

To do - perhaps simplify, or get rid of the /config/manifest.git entirely,
keeping our confguration purely in zot.  We need some place to store uid
mappings, though.

Signed-off-by: Serge E. Hallyn <serge@hallyn.com>