CI: update livecd testscases
The livecd testcases were using the distributed bootkit to install
shim and kernel.  That can only work when we are using the same
keyset, which we no longer are.  Furthermore, bootkit layer no
longer ships kernel.efi at all, or signed shim.  So use the
bootkit from the keyset.

Also, drop the bootkit layer from livecd1.bats and from the
provisioning iso in livecd2.bats, as they're not needed
there.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>