Override mosctl in snakeoil kernel when testing
Add a hidden --mosctl-path flag to 'trust keyset add'.  If set, then an
extra cpio is added to initrd which sets /usr/bin/mosctl to the provided
path.

Use this flag during testing, so that we can pass tests with a new mosctl
even if bootkit shipped with a different mosctl.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>