mosb publish: auto-fill bootkit and hostfs layers
If bootkit layer is not specified, then add the one for the
specified trust org.

If hostfs layer is not specified, use the upstream demo one.

This makes less paperwork for the user.

Add a '--skip-boot' option to mosb manifest publish, for use when we
are testing fake mosctl install.

.github/workflows/build.yml: fetch the git tags, because mosb
must know which tag to use when filling in a rootfs entry.

Signed-off-by: Serge Hallyn <serge@hallyn.com>