implement persistent storage for services
The manifest.yaml now has a storage section where requests for
additional named partitions can be added, e.g.

storage:
  - label: zot-data
    persistent: true
    nsgroup: "zot"
    size: 30G

Each target also has a storage section specifying a list of
labels to mount and destinations at which to mount them.  A
nonpersistent storage label can be used if some target expects
to use a lot of temporary space with which it doesn't want to fill
up the shared /scratch-writes.  Persistent volumes can be used
to maintain data across reboots.

Signed-off-by: Serge Hallyn <serge@hallyn.com>